refactor: unify BaseAgent/Agent/AsyncAgent into agent.py
