+++ python/mirascope/llm/agents/__init__.py
@@ -1,8 +1,6 @@
 """The Agents module for creating and managing LLM agents."""
 
-from .agent import Agent
-from .async_agent import AsyncAgent
-from .base_agent import BaseAgent
+from .agent import Agent, AsyncAgent, BaseAgent
 from .decorator import AgentDecorator, agent
 
 __all__ = [
